pyproject.toml
[project]
name = "mcp-examples"
version = "0.1.0"
description = "Example susing MCP"
authors = [
    {name = "Andres Felipe Tellez"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "fastmcp (>=2.3.3,<3.0.0)",
    "langchain (>=0.3.25,<0.4.0)",
    "langchain-mcp-adapters (>=0.0.11,<0.0.12)",
    "anthropic (>=0.51.0,<0.52.0)",
    "openai (>=1.78.1,<2.0.0)",
    "langchain-openai (>=0.3.16,<0.4.0)",
    "langchain-ollama (>=0.3.2,<0.4.0)",
    "black (>=25.1.0,<26.0.0)",
    "langchain-community (>=0.3.24,<0.4.0)",
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "langgraph (>=0.4.3,<0.5.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<2.2.0"]
build-backend = "poetry.core.masonry.api"
